refactoring DAO method: removeUserById()

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoJDBCImpl.java
@@ -50,13 +50,16 @@ public class UserDaoJDBCImpl implements UserDao {
     }
 
     public void removeUserById(long id) {
-        try (PreparedStatement preparedStatement = connection.prepareStatement(
-                "DELETE FROM users WHERE id = ?")
-        ) {
-            preparedStatement.setLong(1, id);
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
+        if (getUserById(id) != null) {
+            try (PreparedStatement preparedStatement = connection.prepareStatement(
+                    "DELETE FROM users WHERE id = ?")
+            ) {
+                preparedStatement.setLong(1, id);
+                preparedStatement.executeUpdate();
+                System.out.println("INFO: User с id=" + id + " был удалён.");
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
         }
     }
